Adding mroe configs to vim. Configuring bar launch script

+++ .config/polybar/launch.sh
@@ -3,7 +3,8 @@
 ## Add this to your wm startup file.
 
 # Terminate already running bar instances
-if [[ $(command -v waybar) ]]; then
+SESSION_TYPE=$(loginctl show-session $(loginctl| grep '^\s' | awk '{print $1}') -p Type | awk -F= '{print $2}')
+if [[ $(command -v waybar) && $SESSION_TYPE == 'wayland' ]]; then
     killall -q waybar
     export XDG_CURRENT_DESKTOP=Unity
     waybar
